+++ src/app/Http/Controllers/TaxSummaryController.php
@@ -8,22 +8,9 @@ use Illuminate\View\View;
 
 class TaxSummaryController extends Controller
 {
-    public function __invoke(Request $request): View
+    public function __invoke(Request $request, RealizedGainService $realizedGainService): View
     {
-        $user       = $request->user();
-        $service    = new RealizedGainService;
-        $portfolios = $user->portfolios()->where('is_tax_advantaged', false)->with('transactions.asset')->get();
-
-        $allLots = collect();
-
-        foreach ($portfolios as $portfolio) {
-            $result  = $service->compute($portfolio);
-            $allLots = $allLots->concat($result['lots']->map(function ($lot) use ($portfolio) {
-                $lot['portfolio'] = $portfolio;
-
-                return $lot;
-            }));
-        }
+        $allLots = $realizedGainService->allLotsForUser($request->user());
 
         $byYear = $allLots
             ->groupBy(fn ($l) => $l['sell_date']->year)